feat: moved bcrypt and hashing functions to api instead of commons
