+++ apis/auth/package.json
@@ -16,6 +16,7 @@
     "@studio/api-utils": "workspace:packages/api-utils",
     "@studio/commons": "workspace:packages/commons",
     "@studio/dependency-injection": "workspace:packages/dependency-injection",
+    "bcrypt": "^5.1.0",
     "class-transformer": "0.4.0",
     "class-validator": "^0.14.0",
     "dotenv": "^16.0.3",
@@ -34,6 +35,7 @@
   },
   "devDependencies": {
     "@faker-js/faker": "^7.6.0",
+    "@types/bcrypt": "^5.0.0",
     "@types/body-parser": "^1.19.2",
     "@types/express": "^4.17.14",
     "@types/jest": "^29.4.0",
