+++ packages/commons/package.json
@@ -8,13 +8,9 @@
     "build": "rimraf dist && tsc"
   },
   "devDependencies": {
-    "@types/bcrypt": "^5.0.0",
     "@types/node": "^18.11.9",
     "rimraf": "^4.4.1",
     "typescript": "^4.9.3"
   },
-  "packageManager": "yarn@3.3.0",
-  "dependencies": {
-    "bcrypt": "^5.1.0"
-  }
+  "packageManager": "yarn@3.3.0"
 }
